Add render-draw-lines
This commit adds render-draw-lines but I'm not terribly happy with the
implementation. The function will allocate and initialize an SDL_Point *
which then copies the list of SDL_Point structs into.

Presumably one would present their own managed SDL_Point * to this
function which just hands it off to our wrapped FFI call.

However this does "feel" like a more Lisp-y interface. If the extra copy
does prove to be a bottle-neck we may have to revisit this one. I don't
want to establish a policy on how to handle pointer-to-struct until a
pattern emerges.