Add render-draw-rects
This function, like many that take a pointer of struct, has been
implemented by taking a list of struct by the Lisp wrapper which we then
copy into.

    (defun render-draw-foos (render &rest foos))

Appears to be a comfortable Lispy interface and yet the example in this
patch shows how this signature can be rather awkward when the user wants
to manage the collection of structs.