Add *thread-return-channel* for in-main-thread.
I found a program that almost always consistently worked in
windows...and then just stopped for no apparent reason.  The same exact
program would run just fine on Linux.  So I tried CCL instead of
SBCL...and found the same thing.  So I'm guessing that there is some
race condition going on with respect to the way the Windows threading
model is different from Linux that has to do with how the in-main-thread
macro generated a new channel for each use.